Lists: Fix bug for missing end of list
When the end of a list happened on a line that was less than the minimum length, the actual end would not be generated and therefore produced errors in the final `makeinfo` conversion.

By checking is we are still in a list upon reaching a line that is too short, the list is promptly ended.