A sizeable update
If the first line starts with a single `#` followed by a space ` `, then alter title if not specified on command line, and do not create a chapter header from this line. Remove this line from final output.

Also add a flag to hide the table of contents, as Texinfo generates it's own, and having a table of contents inside a table of contents is quite confusing.

I wish python had a `ref` keyword!